# Improve `arrays_equal` performance by eliding bounds check