Hi!

I've copied Cesar Eduardo Barros' code from the [`constant_time_eq` crate](https://github.com/cesarb/constant_time_eq) (CC0).

It basically gives the compiler a hint that bounds checking is not necessary when iterating over the elements of the arrays. The speedup is noticeable, e.g. [here are some benches:](https://github.com/mp4096/rust-constant-time-comp-bench)

```
$ cargo bench

running 6 tests
test tests::bench_functional_generic                       ... bench:     107,932 ns/iter (+/- 1,759)
test tests::bench_functional_specialised                   ... bench:       4,568 ns/iter (+/- 99)
test tests::bench_imperative_generic                       ... bench:     107,690 ns/iter (+/- 1,671)
test tests::bench_imperative_generic_no_bound_checking     ... bench:      73,462 ns/iter (+/- 930)
test tests::bench_imperative_specialised                   ... bench:      72,017 ns/iter (+/- 1,159)
test tests::bench_imperative_specialised_no_bound_checking ... bench:       4,565 ns/iter (+/- 66)

test result: ok. 0 passed; 0 failed; 0 ignored; 6 measured; 0 filtered out
```