@isislovecruft 2017-07-31
Hi @mp4096! This is a pretty sweet hack; I had no idea about this trick. :)

Merged; thanks!